// js/frontend/src/lib/types/ShortLink.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { JsonValue } from "../../../../../libs/entity/bindings/serde_json/JsonValue";

export interface ShortLink {
  id: number;
  short_link: string;
  headers?: JsonValue;
  ip?: string;
  cache_hit: boolean;
  created_at: string;
}